requirements.txt
Django==1.11
MySQL-python==1.2.5
django-filter
slackclient